+++ setty/src/markdown.rs
@@ -17,8 +17,8 @@ pub fn schema_to_markdown(schema: &schemars::Schema) -> String {
 
     write_type(buf, schema["title"].as_str().unwrap(), schema).unwrap();
 
-    for (name, sch) in schema["$defs"].as_object().unwrap() {
-        writeln!(buf).unwrap();
+    let empty = serde_json::Map::default();
+    for (name, sch) in schema["$defs"].as_object().unwrap_or(&empty) {
         writeln!(buf).unwrap();
         write_type(buf, name, sch).unwrap();
     }
@@ -65,8 +65,12 @@ fn write_struct(buf: &mut String, name: &str, schema: &Value) -> Result<(), std:
         .and_then(|p| p.as_object())
         .unwrap_or(&empty);
 
-    writeln!(buf, "| Field | Type | Required | Description |")?;
-    writeln!(buf, "|---|---|---|---|")?;
+    writeln!(buf, "<table>")?;
+    writeln!(
+        buf,
+        "<thead><tr><th>Field</th><th>Type</th><th>Default</th><th>Description</th></tr></thead>"
+    )?;
+    writeln!(buf, "<tbody>")?;
 
     let mut required = std::collections::BTreeSet::new();
     if let Some(req) = schema.get("required").and_then(|v| v.as_array()) {
@@ -78,9 +82,15 @@ fn write_struct(buf: &mut String, name: &str, schema: &Value) -> Result<(), std:
     for (pname, psch) in properties {
         let mut is_required = required.contains(pname.as_str());
 
-        let ptype = if let Some(ty) = psch.get("type") {
+        writeln!(buf, "<tr>")?;
+
+        // Name
+        writeln!(buf, "<td><code>{pname}</code></td>")?;
+
+        // Type
+        if let Some(ty) = psch.get("type") {
             if let Some(ty) = ty.as_str() {
-                format!("`{ty}`")
+                writeln!(buf, "<td><code>{ty}</code></td>")?;
             } else if let Some(ty) = ty.as_array()
                 && ty.len() == 2
             {
@@ -91,33 +101,71 @@ fn write_struct(buf: &mut String, name: &str, schema: &Value) -> Result<(), std:
                     .unwrap();
 
                 is_required = false;
-                format!("`{ty}`")
+                writeln!(buf, "<td><code>{ty}</code></td>")?;
             } else {
                 unreachable!("Unsupported schema in {name}.{pname}:\n{psch}")
             }
         } else if let Some(r) = psch.get("$ref").and_then(|v| v.as_str()) {
             let (_, tname) = r.rsplit_once('/').unwrap();
-            format!("[`{tname}`](#{})", name_to_id(tname))
+            writeln!(
+                buf,
+                "<td><a href=\"#{}\"><code>{tname}</code></a></td>",
+                name_to_id(tname)
+            )?;
         } else if let Some(any_of) = psch.get("anyOf").and_then(|v| v.as_array()) {
             assert_eq!(any_of.len(), 2);
             assert_eq!(any_of[1]["type"].as_str(), Some("null"));
             let r = any_of[0]["$ref"].as_str().unwrap();
             let (_, tname) = r.rsplit_once('/').unwrap();
-            format!("[`{tname}`](#{})", name_to_id(tname))
+            writeln!(
+                buf,
+                "<td><a href=\"#{}\"><code>{tname}</code></a></td>",
+                name_to_id(tname)
+            )?;
         } else {
             unreachable!("Unsupported schema in {name}.{pname}:\n{psch}")
         };
 
-        let pdesc = escape_str_for_table(
-            psch.get("description")
-                .and_then(|v| v.as_str())
-                .unwrap_or(""),
-        );
+        // Default
+        let null = serde_json::Value::Null;
+        if is_required {
+            writeln!(buf, "<td></td>")?;
+        } else {
+            let default = psch.get("default").unwrap_or(&null);
+            let default_str = serde_json::to_string_pretty(default).unwrap();
+            let escaped = html_escape::encode_safe(&default_str);
 
-        let req = if is_required { "`V`" } else { "" };
+            if escaped.contains('\n') {
+                writeln!(
+                    buf,
+                    "<td><pre><code class=\"language-json\">{escaped}</code></pre></td>"
+                )?;
+            } else {
+                writeln!(
+                    buf,
+                    "<td><code class=\"language-json\">{escaped}</code></td>"
+                )?;
+            }
+        };
 
-        writeln!(buf, "| `{pname}` | {ptype} | {req} | {pdesc} |")?;
+        // Description
+        let pdesc = psch
+            .get("description")
+            .and_then(|v| v.as_str())
+            .unwrap_or("")
+            .trim();
+        if pdesc.contains('\n') || pdesc.contains('`') {
+            writeln!(buf, "<td>\n\n{pdesc}\n\n</td>")?;
+        } else {
+            let pdesc = html_escape::encode_script(pdesc);
+            writeln!(buf, "<td>{pdesc}</td>")?;
+        }
+
+        writeln!(buf, "</tr>")?;
     }
+
+    writeln!(buf, "</tbody>")?;
+    writeln!(buf, "</table>")?;
 
     Ok(())
 }
@@ -131,8 +179,9 @@ fn write_enum(buf: &mut String, name: &str, schema: &Value) -> Result<(), std::f
 
     let tag_property = crate::merge_with_defaults::get_enum_tag_property_name(variants);
 
-    writeln!(buf, "| Variants |")?;
-    writeln!(buf, "|---|")?;
+    writeln!(buf, "<table>")?;
+    writeln!(buf, "<thead><tr><th>Variants</th></tr></thead>")?;
+    writeln!(buf, "<tbody>")?;
 
     for variant in variants.iter() {
         let tag = variant["properties"][tag_property]["const"]
@@ -141,8 +190,15 @@ fn write_enum(buf: &mut String, name: &str, schema: &Value) -> Result<(), std::f
 
         let vname = format!("{name}::{tag}");
 
-        writeln!(buf, "| [`{tag}`](#{}) |", name_to_id(&vname))?;
+        writeln!(
+            buf,
+            "<tr><td><a href=\"#{}\"><code>{tag}</code></a></td></tr>",
+            name_to_id(&vname)
+        )?;
     }
+
+    writeln!(buf, "</tbody>")?;
+    writeln!(buf, "</table>")?;
 
     for vsch in variants.iter() {
         writeln!(buf)?;
@@ -159,12 +215,21 @@ fn write_enum(buf: &mut String, name: &str, schema: &Value) -> Result<(), std::f
 /////////////////////////////////////////////////////////////////////////////////////////
 
 fn write_string_enum(buf: &mut String, _name: &str, schema: &Value) -> Result<(), std::fmt::Error> {
-    writeln!(buf, "| Variants |")?;
-    writeln!(buf, "|---|")?;
+    writeln!(buf, "<table>")?;
+    writeln!(buf, "<thead><tr><th>Variants</th></tr></thead>")?;
+    writeln!(buf, "<tbody>")?;
 
     for variant in schema.get("enum").and_then(|v| v.as_array()).unwrap() {
-        writeln!(buf, "| `{}` |", variant.as_str().unwrap())?;
+        let tag = variant.as_str().unwrap();
+        writeln!(
+            buf,
+            "<tr><td><code>{}</code></td></tr>",
+            html_escape::encode_safe(tag)
+        )?;
     }
+
+    writeln!(buf, "</tbody>")?;
+    writeln!(buf, "</table>")?;
 
     Ok(())
 }
@@ -185,16 +250,6 @@ fn write_basic_type_wrapper(
 
 fn name_to_id(name: &str) -> String {
     name.replace(":", "").to_lowercase()
-}
-
-/////////////////////////////////////////////////////////////////////////////////////////
-
-// HACKS: TODO: Proper Markdown parser?
-fn escape_str_for_table(s: &str) -> String {
-    s.replace("```sh", "")
-        .replace("```", "")
-        .replace("|", "\\|")
-        .replace("\n", "<br>")
 }
 
 /////////////////////////////////////////////////////////////////////////////////////////