+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Support validation using `validator` crate
+- Improved markdown generator
 
 ## [0.3.9] - 2026-02-14
 ### Added